Implement an LRU eviction policy for wake's shared cache
